# packages/openapi/admin.yaml
openapi: 3.0.3
info:
  title: Nexus Admin API
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
...
  /admin/oc-days:
    get:
      summary: List open campus days
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OcDay list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OcDayResponse'
    post:
      summary: Create open campus day
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OcDayRequest'
      responses:
        '201':
          description: Created open campus day
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OcDayResponse'
  /admin/oc-days/{id}:
    delete:
      summary: Delete open campus day
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OcDayRequest:
      type: object
      required: [date]
      properties:
        date:
          type: string
          format: date
        name:
          type: string
          nullable: true
    OcDayResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        name:
          type: string
          nullable: true
...